+++ src/Microsoft.ApplicationInsights.AspNetCore/RequestTrackingTelemetryModule.cs
@@ -8,6 +8,7 @@ namespace Microsoft.ApplicationInsights.AspNetCore
     using Microsoft.ApplicationInsights.AspNetCore.DiagnosticListeners;
     using Microsoft.ApplicationInsights.AspNetCore.Extensibility.Implementation.Tracing;
     using Microsoft.ApplicationInsights.AspNetCore.Extensions;
+    using Microsoft.ApplicationInsights.AspNetCore.Implementation;
     using Microsoft.ApplicationInsights.Extensibility;
     using Microsoft.AspNetCore.Hosting;
 
@@ -68,14 +69,28 @@ namespace Microsoft.ApplicationInsights.AspNetCore
                         {
                             this.telemetryClient = new TelemetryClient(configuration);
 
-                            bool enableNewDiagnosticEvents = true;
+                            AspNetCoreMajorVersion aspNetCoreMajorVersion = AspNetCoreMajorVersion.Two;
+
                             try
                             {
-                                enableNewDiagnosticEvents = typeof(IWebHostBuilder).GetTypeInfo().Assembly.GetName().Version.Major >= 2;
+                                var version = typeof(IWebHostBuilder).GetTypeInfo().Assembly.GetName().Version.Major;
+
+                                if (version < 2)
+                                {
+                                    aspNetCoreMajorVersion = AspNetCoreMajorVersion.One;
+                                }
+                                else if (version == 2)
+                                {
+                                    aspNetCoreMajorVersion = AspNetCoreMajorVersion.Two;
+                                }
+                                else
+                                {
+                                    aspNetCoreMajorVersion = AspNetCoreMajorVersion.Three;
+                                }                                
                             }
-                            catch (Exception)
+                            catch (Exception e)
                             {
-                                // ignore any errors
+                                AspNetCoreEventSource.Instance.LogError($"Exception occured while attempting to find Asp.Net Core Major version. Assuming {aspNetCoreMajorVersion.ToString()} and continuing. Exception: {e.Message}");
                             }
 
                             this.diagnosticListener = new HostingDiagnosticListener(
@@ -85,12 +100,12 @@ namespace Microsoft.ApplicationInsights.AspNetCore
                                 this.CollectionOptions.InjectResponseHeaders,
                                 this.CollectionOptions.TrackExceptions,
                                 this.CollectionOptions.EnableW3CDistributedTracing,
-                                enableNewDiagnosticEvents);
+                                aspNetCoreMajorVersion);
 
                             this.subscriptions?.Add(DiagnosticListener.AllListeners.Subscribe(this));
 
                             // Questionable to modify the configuration here.
-                            configuration.EnableW3CCorrelation = this.CollectionOptions.EnableW3CDistributedTracing;
+                             configuration.EnableW3CCorrelation = this.CollectionOptions.EnableW3CDistributedTracing;
                             this.isInitialized = true;
                         }
                     }